npm compatibility, console testability & pre-generated parser
1. npm compatibility

The core filtrex library has been modified so it can be installed and require()’d by node.js projects. The previous primary build artifact `filtrex.js` (and new filtrex.min.js) remains as a UMD-compatible module, generated by browserify (and uglify-js).

2. Console testability

The test suite has been minimally modified to allow the running of tests via `npm test`.

3. Pre-generated parser

To improve runtime efficiency and file-size, the Jison parser is precompiled into src/parser.js.
