+++ src/grammar.json
@@ -1,0 +1,1 @@
+{"lex":{"rules":[["\\*","return \"*\";"],["\\/","return \"/\";"],["-","return \"-\";"],["\\+","return \"+\";"],["\\^","return \"^\";"],["\\%","return \"%\";"],["\\(","return \"(\";"],["\\)","return \")\";"],["\\,","return \",\";"],["==","return \"==\";"],["\\!=","return \"!=\";"],[">=","return \">=\";"],["<=","return \"<=\";"],["<","return \"<\";"],[">","return \">\";"],["\\?","return \"?\";"],["\\:","return \":\";"],["and[^\\w]","return \"and\";"],["or[^\\w]","return \"or\";"],["not[^\\w]","return \"not\";"],["in[^\\w]","return \"in\";"],["\\s+",""],["[0-9]+(?:\\.[0-9]+)?\\b","return \"NUMBER\";"],["[a-zA-Z][\\.a-zA-Z0-9_]*","return \"SYMBOL\";"],["\"(?:[^\"])*\"","yytext = yytext.substr(1, yyleng-2); return \"STRING\";"],["$","return \"EOF\";"]]},"operators":[["left","?",":"],["left","or"],["left","and"],["left","in"],["left","==","!="],["left","<","<=",">",">="],["left","+","-"],["left","*","/","%"],["left","^"],["left","not"],["left","UMINUS"]],"bnf":{"expressions":[["e EOF","return $1;"]],"e":[["e + e","$$ = [\"(\", $1,\"+\",$3, \")\"];"],["e - e","$$ = [\"(\", $1,\"-\",$3, \")\"];"],["e * e","$$ = [\"(\", $1,\"*\",$3, \")\"];"],["e / e","$$ = [\"(\", $1,\"/\",$3, \")\"];"],["e % e","$$ = [\"(\", $1,\"%\",$3, \")\"];"],["e ^ e","$$ = [\"(\", \"Math.pow(\",$1,\",\",$3,\")\", \")\"];"],["- e","$$ = [\"(\", \"-\",$2, \")\"];",{"prec":"UMINUS"}],["e and e","$$ = [\"(\", \"Number(\",$1,\"&&\",$3,\")\", \")\"];"],["e or e","$$ = [\"(\", \"Number(\",$1,\"||\",$3,\")\", \")\"];"],["not e","$$ = [\"(\", \"Number(!\",$2,\")\", \")\"];"],["e == e","$$ = [\"(\", \"Number(\",$1,\"==\",$3,\")\", \")\"];"],["e != e","$$ = [\"(\", \"Number(\",$1,\"!=\",$3,\")\", \")\"];"],["e < e","$$ = [\"(\", \"Number(\",$1,\"<\",$3,\")\", \")\"];"],["e <= e","$$ = [\"(\", \"Number(\",$1,\"<=\",$3,\")\", \")\"];"],["e > e","$$ = [\"(\", \"Number(\",$1,\"> \",$3,\")\", \")\"];"],["e >= e","$$ = [\"(\", \"Number(\",$1,\">=\",$3,\")\", \")\"];"],["e ? e : e","$$ = [\"(\", $1,\"?\",$3,\":\",$5, \")\"];"],["( e )","$$ = [\"(\", $2, \")\"];"],["NUMBER","$$ = [\"(\", $1, \")\"];"],["STRING","$$ = [\"(\", \"\\\"\",$1,\"\\\"\", \")\"];"],["SYMBOL","$$ = [\"(\", \"data[\\\"\",$1,\"\\\"]\", \")\"];"],["SYMBOL ( argsList )","$$ = [\"(\", \"functions.\",$1,\"(\",$3,\")\", \")\"];"],["e in ( inSet )","$$ = [\"(\", $1,\" in (function(o) { \",$4,\"return o; })({})\", \")\"];"],["e not in ( inSet )","$$ = [\"(\", \"!(\",$1,\" in (function(o) { \",$5,\"return o; })({}))\", \")\"];"]],"argsList":[["e","$$ = [$1];"],["argsList , e","$$ = [$1,\",\",$3];"]],"inSet":[["e","$$ = [\"o[\",$1,\"] = true; \"];"],["inSet , e","$$ = [$1,\"o[\",$3,\"] = true; \"];"]]}}
